feat(list): add item management features
- Implement ListItemLeftActions for item check/uncheck functionality.
- Implement ListItemRightActions for item edit and delete actions.
- Create ListItemSkeleton for loading states.
- Add ListItemsContent to render list items with sorting and filtering.
- Create ListItemsEmptyComponent for empty state display.
- Implement ListItemsFooter to show total amount.
- Add ListItemsSortBar for sorting options.
- Create modals for item creation, updating, and deletion.
- Implement useListItemsPageLogics for managing list item state and actions.
- Add utility functions for currency formatting and validation.
- Implement sorting utilities for list items.

This change enhances the list management experience by providing a complete set of actions for item manipulation, improving user interaction and data handling.
