--- src/features/list-items/index.ts
@@ -1,10 +1,0 @@
-/**
- * Barrel export para a feature list
- */
-
-export { default as ListItemsScreen } from './page';
-export * from './types';
-export * from './utils';
-export * from './components';
-export * from './modals';
-export * from './hooks/use-list-items-page-logics';
