+++ src/features/list/utils/validation.ts
